* ext/dl/cfunc.c (rb_dlcfunc_inspect): get rid of overflow.
* ext/dl/closure.c (dlc_callback, rb_dlclosure_init): ditto.

* ext/dl/cptr.c (rb_dlptr_s_malloc): ditto.

* ext/dl/method.c (rb_dlfunction_initialize): ditto.


git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@26693 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/dl/cfunc.c
@@ -284,7 +284,7 @@ rb_dlcfunc_inspect(VALUE self)
 {
     VALUE val;
     char  *str;
-    int str_size;
+    size_t str_size;
     struct cfunc_data *cfunc;
 
     TypedData_Get_Struct(self, struct cfunc_data, &dlcfunc_data_type, cfunc);
